memory/requirements.txt
# Grand Prix Social Memory System Dependencies
# Core dependencies for the cognitive memory system

# Core Python packages
numpy>=1.21.0
scipy>=1.8.0
pandas>=1.4.0

# AI and ML
openai>=1.0.0
anthropic>=0.8.0
transformers>=4.20.0

# Web and API
requests>=2.28.0
httpx>=0.24.0
aiohttp>=3.8.0
fastapi>=0.68.0
uvicorn>=0.15.0

# Database and Storage
redis>=4.0.0
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0

# File handling and monitoring
watchdog>=2.1.0
python-dotenv>=0.19.0
PyYAML>=6.0

# System utilities
psutil>=3.4.0
click>=8.0.0
rich>=12.0.0
typer>=0.4.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Async and concurrency
asyncio-mqtt>=0.11.0
celery>=5.2.0

# Data processing
markdown>=3.3.0
Pillow>=9.0.0

# Logging and monitoring
structlog>=22.1.0
prometheus-client>=0.14.0

# Optional AI packages (for advanced features)
sentence-transformers>=2.2.0
torch>=1.12.0
scikit-learn>=1.1.0

# Memory and caching
diskcache>=5.4.0
joblib>=1.1.0